% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emboss.R
\name{emboss}
\alias{emboss}
\title{This function embosses the original JPG image and saves the embossed image to out_path.}
\usage{
emboss(input_path, output_path)
}
\arguments{
\item{input_path}{(character) -> the file path for the original image we want to emboss}

\item{output_path}{(character) ->  the file path to save the embossed image}
}
\value{
No return. Transformed image will be saved into the output_path
}
\description{
This function embosses the original JPG image and saves the embossed image to out_path.
}
