% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip}
\usage{
flip(input_path, output_path, direction)
}
\arguments{
\item{input_path}{string, path for the input png file}

\item{output_path}{string, path for the output png file}

\item{direction}{string, direction of flip}
}
\value{
a png file at the output path
}
\description{
Flip
}
\examples{
flip("test_imgs/flip/test_input.png","test_imgs/flip/test_output.png","v")
}
