% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_details.R
\name{get_image_details}
\alias{get_image_details}
\title{This function returns details about the image, such as dimension, width, height, and image ratio.}
\usage{
get_image_details(input_path, detail = "All")
}
\arguments{
\item{input_path}{(string) ->  The file path for the image we want to return information of.}

\item{detail}{(string) -> The name of attribute that the function will return. Default set to be 'All'. Available choices are: 'All', 'Dimension', 'Width', 'Height', and 'Aspect Ratio'.}
}
\value{
A data frame that has the detailed information about input image.
}
\description{
This function returns details about the image, such as dimension, width, height, and image ratio.
}
